backend/requirements.txt
fastapi
uvicorn
mangum
boto3
pydantic
qdrant-client
google-generativeai
